final adjustments and updated deps
