+++ src/configs/types/config_settings.rs
@@ -25,12 +25,6 @@ impl ConfigSettings {
         if let Ok(settings) = confy::load::<ConfigSettings>("sniffnet", "settings") {
             settings
         } else {
-            println!(
-                "{:?}",
-                confy::load::<ConfigSettings>("sniffnet", "settings")
-                    .err()
-                    .unwrap()
-            );
             confy::store("sniffnet", "settings", ConfigSettings::default()).unwrap_or(());
             ConfigSettings::default()
         }
